fix: 清理 C/C++ 编译警告 + 补充缺失错误码
- hook_fileapi.c: C4244 uint64→uint32 截断警告修复
- loader_main.c: C4146 无符号取负警告修复（0u - (crc & 1)）
- lzma_dec.c: 删除重复/格式错误注释行
- loader_errors.h: 新增 VFS_ERR_NO_MEMORY (6009) 错误码

+++ src/EniBox.Loader/src/hook_fileapi.c
@@ -120,10 +120,10 @@ DWORD WINAPI Hook_SetFilePointer(HANDLE hFile, LONG lDistanceToMove,
                 new_pos = (uint32_t)lDistanceToMove;
                 break;
             case FILE_CURRENT:
-                new_pos = h->current_pos + (uint32_t)lDistanceToMove;
+                new_pos = (uint32_t)((uint64_t)h->current_pos + (int64_t)lDistanceToMove);
                 break;
             case FILE_END:
-                new_pos = file_size + (uint32_t)lDistanceToMove;
+                new_pos = (uint32_t)((uint64_t)file_size + (int64_t)lDistanceToMove);
                 break;
             default:
                 SetLastError(ERROR_INVALID_PARAMETER);
